worker_cej/requirements.txt
requests==2.32.3
aio-pika==9.5.6
dependency-injector==4.45.0
pydantic-settings==2.6.1
oracledb==3.1.1
pdfminer.six>=20201018
boto3
botocore==1.35.36
aiohttp
PyMuPDF==1.24.3
openpyxl
pandas
pypandoc
lxml
beautifulsoup4
python-magic
user-agent